+++ lib/DCS-BIOS/src/internal/Protocol.h
@@ -9,15 +9,16 @@
 #define DCSBIOS_STATE_DATA_LOW 5
 #define DCSBIOS_STATE_DATA_HIGH 6
 
+#include <stdint.h>
 #include "ExportStreamListener.h"
 
 namespace DcsBios {
 	class ProtocolParser {
 	private:
 		volatile unsigned char state;
-		volatile unsigned int address;
-		volatile unsigned int count;
-		volatile unsigned int data;
+		volatile uint16_t address;
+		volatile uint16_t count;
+		volatile uint16_t data;
 		volatile unsigned char sync_byte_count;
 		ExportStreamListener* startESL;
 		volatile bool processingData;
